Let simulated devices know the simulated system time.
